Uppdate int to double

+++ Line.cs
@@ -8,9 +8,9 @@ namespace Simulateur_Réseau
 	{
 		public entrance; //peut etre ajouter noeud dans les acteur et rename la classe en lieu de deplacement elec.
 		public exit;
-		public int powerMax;
-		public int powerIn;
-		public int powerNeeded;
+		public double powerMax;
+		public double powerIn;
+		public double powerNeeded;
 
 		public Line(Actor incoming, )
 		{
@@ -19,17 +19,17 @@ namespace Simulateur_Réseau
 			this.exit = //
 		}
 
-		public setPowerIn(int powerIn)
+		public setPowerIn(double powerIn)
 		{
 			this.powerIn = powerIn;
 		}
 
-		public int getPowerIn()
+		public double getPowerIn()
 		{
 			return this.powerIn;
 		}
 
-		public setPowerNeeded(int powerNeeded)
+		public setPowerNeeded(double powerNeeded)
 		{
 			if (powerNeeded < powerMax)
 			{
@@ -40,7 +40,7 @@ namespace Simulateur_Réseau
 				return; //message d'erreur cette ligne ne peut pas fournir cette énergie.
 			}
 		}
-		public int getPowerNeeded()
+		public double getPowerNeeded()
 		{
 			return this.powerNeeded;
 		}
